sme_doc_extract_local/requirements.txt
google-cloud-documentai>=2.20.0
google-generativeai>=0.5.0
pydantic>=2.5.0
python-dateutil>=2.8.2
rich>=13.7.0
python-dotenv>=1.0.0
